Organize file folders, update imports
