+++ app/src/UI/Features/Records/Activity/forms/common/CheckboxUI/checkboxUI.css
